Move strings.Dereference to ptrs.Deref
